fix:format code with prettier
