+++ .github/pull_request_template.md
@@ -1,12 +1,20 @@
 you can check this one that was used for the previous cohort
+
 # Fixes Issue/Linear Ticket
+
 > This could be an existing issue or a linear ticket
+>
 > - Github Issue Example: My PR Closes #{ISSUE}
 > - Linear Ticket Example: Fixes ID-#{ISSUE}
+
 # Changes proposed
+
 > Talk about the things you did eg. files changes, dependencies installed e.t.c
+
 # Check List (Check all the applicable boxes)
+
 :rotating_light:Please review the [style guide for contributing](add link here) and [guidelines for contributing](add link here) to this repository.
+
 - [ ] My code follows the code style of this project.
 - [ ] This PR does not contain plagiarized content.
 - [ ] The title and description of the PR is clear and explains the approach.
@@ -14,7 +22,9 @@ you can check this one that was used for the previous cohort
 - [ ] My commit messages styles matches our requested structure.
 - [ ] My code additions will fail neither code linting checks nor unit test.
 - [ ] I am only making changes to files I was requested to.
+
 # Screenshots/ Videos
+
 <!-- If the changes are static page changes or UI changes add screenshots -->
 <!-- If the changes involve implementing a functionality or working with apis, include a video
 detailing how to implement the functionality and the request to the api and responses from the api endpoint-->
